requirements.txt
streamlit
langchain
groq
langchain_groq
python-dotenv